Correction du code Modification de clamp() et dépréciation Nouvel ennemi : EnemyBoss

+++ WaveGame/src/GameObject.java
@@ -1,11 +1,11 @@
 import java.awt.*;
 
 public abstract class GameObject {
-    protected int x, y;
-    protected int velX, velY;
+    protected float x, y;
+    protected double velX, velY;
     protected ID id;
 
-    public GameObject(int x, int y, ID id) {
+    public GameObject(float x, float y, ID id) {
         this.x = x;
         this.y = y;
         this.id = id;
@@ -14,28 +14,28 @@ public abstract class GameObject {
     public abstract void tick();
     public abstract void render(Graphics g);
 
-    public int getX() {
+    public float getX() {
         return x;
     }
     public void setX(int x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
     public void setY(int y) {
         this.y = y;
     }
 
-    public int getVelX() {
+    public double getVelX() {
         return velX;
     }
     public void setVelX(int velX) {
         this.velX = velX;
     }
 
-    public int getVelY() {
+    public double getVelY() {
         return velY;
     }
     public void setVelY(int velY) {
